@@ -16,40 +16,43 @@
   "license": "MIT",
   "homepage": "https://github.com/trambarhq/trambar-deco",
   "devDependencies": {
-    "babel-core": "^6.24.1",
-    "babel-loader": "^7.0.0",
-    "babel-preset-env": "^1.4.0",
-    "babel-preset-es2015": "^6.24.1",
-    "babel-preset-react": "^6.24.1",
-    "css-loader": "^0.28.1",
-    "file-loader": "^0.11.1",
+    "@babel/core": "^7.8.7",
+    "@babel/plugin-transform-runtime": "^7.8.3",
+    "@babel/preset-env": "^7.8.7",
+    "@babel/preset-react": "^7.8.3",
+    "@babel/runtime": "^7.8.7",
+    "babel-loader": "^8.0.6",
+    "css-loader": "^3.4.2",
+    "file-loader": "^5.1.0",
     "font-awesome": "^4.7.0",
-    "font-awesome-webpack": "0.0.5-beta.2",
-    "html-loader": "^0.4.5",
-    "html-webpack-plugin": "^2.28.0",
-    "less": "^2.7.2",
-    "node-sass": "^4.5.3",
-    "react": "~15.4.0",
-    "react-dom": "~15.4.0",
-    "sass-loader": "^6.0.5",
-    "sockjs-client": "^1.1.4",
-    "style-loader": "^0.17.0",
-    "uglifyjs-webpack-plugin": "^0.4.6",
-    "url-loader": "^0.5.8",
-    "webpack": "^2.5.1",
-    "webpack-dev-server": "^2.4.5"
+    "html-loader": "^0.5.5",
+    "html-webpack-plugin": "^3.2.0",
+    "less": "^3.11.1",
+    "node-sass": "^4.13.1",
+    "prop-types": "^15.7.2",
+    "react": "~16.13.0",
+    "react-dom": "~16.13.0",
+    "relaks": "^2.1.11",
+    "sass-loader": "^8.0.2",
+    "sockjs-client": "^1.4.0",
+    "style-loader": "^1.1.3",
+    "uglifyjs-webpack-plugin": "^2.2.0",
+    "url-loader": "^3.0.0",
+    "webpack": "^4.42.0",
+    "webpack-cli": "^3.3.11",
+    "webpack-dev-server": "^3.10.3"
   },
   "dependencies": {
-    "bluebird": "^3.5.0",
-    "body-parser": "^1.17.2",
-    "chokidar": "^1.7.0",
-    "command-line-args": "^4.0.7",
-    "command-line-usage": "^4.0.1",
-    "express": "^4.15.4",
-    "ignore": "^3.3.5",
-    "lodash": "^4.17.4",
-    "mark-gor": "0.0.6",
-    "open": "0.0.5",
-    "sockjs": "^0.3.18"
+    "bluebird": "^3.7.2",
+    "body-parser": "^1.19.0",
+    "chokidar": "^3.3.1",
+    "command-line-args": "^5.1.1",
+    "command-line-usage": "^6.1.0",
+    "express": "^4.17.1",
+    "ignore": "^5.1.4",
+    "lodash": "^4.17.15",
+    "mark-gor": "2.1.1",
+    "open": "7.0.3",
+    "sockjs": "^0.3.20"
   }
 }
