@@ -1,15 +1,11 @@
-var React = require('react');
-var ReactDOM = require('react-dom');
-
-var Application = require('application');
+import { createElement } from 'react';
+import { render } from 'react-dom';
+import { Application } from './application.jsx';
 
 window.addEventListener('load', initialize);
 
 function initialize(evt) {
-    var appContainer = document.getElementById('app-container');
-    if (!appContainer) {
-        throw new Error('Unable to find app element in DOM');
-    }
-    var appElement = React.createElement(Application);
-    ReactDOM.render(appElement, appContainer);
+  const appContainer = document.getElementById('react-container');
+  const appElement = createElement(Application);
+  render(appElement, appContainer);
 }
