@@ -1,163 +1,116 @@
-var Promise = require('bluebird');
-var React = require('react'), PropTypes = React.PropTypes;
+import React, { useState, useRef } from 'react';
+import { useListener } from 'relaks';
 
 // widgets
-var Overlay = require('widgets/overlay');
-var PushButton = require('widgets/push-button');
+import { Overlay } from './overlay.jsx';
+import { PushButton } from './push-button.jsx';
 
-require('./font-awesome-dialog-box.scss');
+import './font-awesome-dialog-box.scss';
 
-module.exports = React.createClass({
-    displayName: 'FontAwesomeDialogBox',
-    propTypes: {
-        show: PropTypes.bool,
-        iconClassName: PropTypes.string,
-        onClose: PropTypes.func,
-    },
+export const FontAwesomeDialogBox = Overlay.create((props) => {
+  const { show, iconClassName, onClose } = props;
+  const [ color, setColor ] = useState(() => {
+    return localStorage.iconColor || '#000000';
+  });
+  const [ backgroundColor, setBackgroundColor ] = useState(() => {
+    return localStorage.iconBackgroundColor || '#fca326';
+  });
+  const urlInputRef = useRef();
 
-    /**
-     * Return initial state of component
-     *
-     * @return {Object}
-     */
-    getInitialState: function() {
-        return {
-            color: localStorage.iconColor || '#000000',
-            backgroundColor: localStorage.iconBackgroundColor || '#fca326',
-        };
-    },
-
-    /**
-     * Render component
-     *
-     * @return {ReactElement}
-     */
-    render: function() {
-        if (!this.props.iconClassName) {
-            return null;
-        }
-        var overlayProps = {
-            show: this.props.show,
-            onBackgroundClick: this.props.onClose,
-        };
-        return (
-            <Overlay {...overlayProps}>
-                <div className="font-awesome-dialog-box">
-                    {this.renderForm()}
-                    {this.renderButtons()}
-                </div>
-            </Overlay>
-        );
-    },
-
-    /**
-     * Render form
-     *
-     * @return {ReactElement}
-     */
-    renderForm: function() {
-        var color = this.state.color;
-        var backgroundColor = this.state.backgroundColor;
-        var url = `fa://${this.props.iconClassName}/${backgroundColor}`;
-        if (color !== '#000000') {
-            url += `/${color}`;
-        }
-        var urlRef = (node) => { this.urlInput = node };
-        return (
-            <div className="form">
-                {this.renderIcon()}
-                <div className="input">
-                    <label>Foreground color</label>
-                    <input type="color" value={color} onChange={this.handleColorChange} />
-                </div>
-                <div className="input">
-                    <label>Background color</label>
-                    <input type="color" value={backgroundColor} onChange={this.handleBackgroundColorChange} />
-                </div>
-                <div className="input">
-                    <label>URL</label>
-                    <input type="text" ref={urlRef} value={url} readOnly={true} />
-                </div>
-            </div>
-        );
-    },
-
-    /**
-     * Render icon
-     *
-     * @return {ReactElement}
-     */
-    renderIcon: function() {
-        var className = this.props.iconClassName;
-        var style = {
-            color: this.state.color,
-            backgroundColor: this.state.backgroundColor,
-        };
-        return (
-            <div className="icon" style={style}>
-                <i className={`fa ${className} fa-fw`} />
-            </div>
-        );
-    },
-
-    /**
-     * Render buttons
-     *
-     * @return {ReactElement}
-     */
-    renderButtons: function() {
-        var cancelButtonProps = {
-            label: 'Cancel',
-            onClick: this.props.onClose,
-        };
-        var copyButtonProps = {
-            label: 'Copy',
-            emphasized: true,
-            onClick: this.handleCopyClick,
-        };
-        return (
-            <div className="buttons">
-                <PushButton {...cancelButtonProps} />
-                <PushButton {...copyButtonProps} />
-            </div>
-        );
-    },
-
-    /**
-     * Called when user changes the color
-     *
-     * @param  {Event} evt
-     */
-    handleColorChange: function(evt) {
-        var color = evt.target.value;
-        this.setState({ color });
-        localStorage.iconColor = color;
-    },
-
-    /**
-     * Called when user changes the background color
-     *
-     * @param  {Event} evt
-     */
-    handleBackgroundColorChange: function(evt) {
-        var color = evt.target.value;
-        this.setState({ backgroundColor: color });
-        localStorage.iconBackgroundColor = color;
-    },
-
-    /**
-     * Called when user clicks copy button
-     *
-     * @param  {Event} evt
-     */
-    handleCopyClick: function(evt) {
-        var node = this.urlInput;
-        if (node) {
-            node.select();
-            document.execCommand('copy');
-        }
-        if (this.props.onClose) {
-            this.props.onClose(evt);
-        }
+  const handleColorChange = useListener((evt) => {
+    const { value } = evt.target;
+    setColor(value);
+    localStorage.iconColor = value;
+  });
+  const handleBackgroundColorChange = useListener((evt) => {
+    const { value } = evt.target;
+    setBackgroundColor(value);
+    localStorage.iconBackgroundColor = value;
+  });
+  const handleCopyClick = useListener((evt) => {
+    const node = urlInputRef.current;
+    if (node) {
+      node.select();
+      document.execCommand('copy');
+      if (onClose) {
+        onClose();
+      }
     }
+  });
+
+  if (!iconClassName) {
+    return null;
+  }
+  return (
+    <div className="font-awesome-dialog-box">
+      {renderForm()}
+      {renderButtons()}
+    </div>
+  );
+
+  /**
+   * Render form
+   *
+   * @return {ReactElement}
+   */
+  function renderForm() {
+    let url = `fa://${iconClassName}/${backgroundColor}`;
+    if (color !== '#000000') {
+      url += `/${color}`;
+    }
+    return (
+      <div className="form">
+        {renderIcon()}
+        <div className="input">
+          <label>Foreground color</label>
+          <input type="color" value={color} onChange={handleColorChange} />
+        </div>
+        <div className="input">
+          <label>Background color</label>
+          <input type="color" value={backgroundColor} onChange={handleBackgroundColorChange} />
+        </div>
+        <div className="input">
+          <label>URL</label>
+          <input type="text" ref={urlInputRef} value={url} readOnly={true} />
+        </div>
+      </div>
+    );
+  }
+
+  /**
+   * Render icon
+   *
+   * @return {ReactElement}
+   */
+  function renderIcon() {
+    const style = { color, backgroundColor };
+    return (
+      <div className="icon" style={style}>
+        <i className={`fa ${iconClassName} fa-fw`} />
+      </div>
+    );
+  }
+
+  /**
+   * Render buttons
+   *
+   * @return {ReactElement}
+   */
+  function renderButtons() {
+    const cancelButtonProps = {
+      label: 'Cancel',
+      onClick: onClose,
+    };
+    const copyButtonProps = {
+      label: 'Copy',
+      emphasized: true,
+      onClick: handleCopyClick,
+    };
+    return (
+      <div className="buttons">
+        <PushButton {...cancelButtonProps} />
+        <PushButton {...copyButtonProps} />
+      </div>
+    );
+  }
 });
