@@ -1,61 +1,46 @@
-var _ = require('lodash');
-var React = require('react'), PropTypes = React.PropTypes;
+import React from 'react';
 
-require('./tree-node-file.scss');
+import './tree-node-file.scss';
 
-module.exports = TreeNodeFile;
+export function TreeNodeFile(props) {
+  const { file, onSelect } = props;
+  const { components = [] } = file;
+  let icon = 'file-o';
+  if (file.text) {
+    icon = 'file-text-o';
+  } else if (/\.(jpe?g|png|gif|svg|bmp|psd)$/i.test(file.path)) {
+    icon = 'file-image-o';
+  } else if (/\.pdf$/i.test(file.path)) {
+    icon = 'file-pdf-o';
+  }
+  const classNames = [ 'tree-node-file' ];
+  if (components.length > 0) {
+    classNames.push('decorated');
+  }
+  return (
+    <div className={classNames.join(' ')}>
+      <i className={`fa fa-fw fa-${icon}`} />
+      {file.path}
+      {renderComponentBadges()}
+    </div>
+  );
 
-function TreeNodeFile(props) {
-    var icon = 'file-o';
-    var file = props.file;
-    if (file.text) {
-        icon = 'file-text-o';
-    } else if (/\.(jpe?g|png|gif|svg|bmp|psd)$/i.test(file.path)) {
-        icon = 'file-image-o';
-    } else if (/\.pdf$/i.test(file.path)) {
-        icon = 'file-pdf-o';
+  function renderComponentBadges() {
+    if (!onSelect) {
+      return null;
     }
-    var className = 'tree-node-file';
-    if (!_.isEmpty(file.components)) {
-        className += ' decorated';
-    }
-    return (
-        <div className={className}>
-            <i className={`fa fa-fw fa-${icon}`} />
-            {file.path}
-            {renderComponentBadges(props)}
+    const { components = [] } = file;
+    const badges = components.map((id, index) => {
+      const handleClick = (evt) => { onSelect({ id }) };
+      return (
+        <div key={index} className="badge" onClick={handleClick}>
+          {index + 1}
         </div>
-    );
-}
-
-TreeNodeFile.propTypes = {
-    file: PropTypes.object.isRequired,
-    onSelect: PropTypes.func,
-};
-
-function renderComponentBadges(props) {
-    if (!props.onSelect) {
-        return null;
-    }
-    var componentIds = props.file.components;
-    if (_.isEmpty(componentIds)) {
-        return null;
-    }
-    var badges = _.map(componentIds, (id, index) => {
-        var handleClick = (evt) => {
-            if (props.onSelect) {
-                props.onSelect({
-                    type: 'select',
-                    target: evt.target,
-                    id,
-                })
-            }
-        };
-        return (
-            <div key={index} className="badge" onClick={handleClick}>
-                {index + 1}
-            </div>
-        );
+      );
     });
+    if (badges.length === 0) {
+      return null;
+    }
     return <div className="badges">{badges}</div>;
+  }
 }
