@@ -1,69 +1,53 @@
-var _ = require('lodash');
-var React = require('react'), PropTypes = React.PropTypes;
+import React from 'react';
 
-var TreeNodeFile = require('widgets/tree-node-file');
+import { TreeNodeFile } from './tree-node-file.jsx';
 
-require('./tree-node-folder.scss');
+import './tree-node-folder.scss';
 
-var TreeNodeFolder = module.exports = React.createClass({
-    displayName: 'TreeNodeFolder',
-    propTypes: {
-        folder: PropTypes.object,
-        root: PropTypes.string,
-        onSelect: PropTypes.func,
-    },
+export function TreeNodeFolder(props) {
+  const { folder, root, onSelect } = props;
+  const label = folder.path || '[ROOT]';
+  return (
+    <div className="tree-node-folder">
+      <i className="fa fa-folder fa-fw" />
+      {label} {renderWorkingFolder()}
+      <div className="files">
+        {folder.children.map(renderChild)}
+      </div>
+    </div>
+  );
 
-    /**
-     * Render component
-     *
-     * @return {ReactElement}
-     */
-    render: function() {
-        var folder = this.props.folder;
-        if (!folder) {
-            return null;
-        }
-        var label = folder.path || '[ROOT]';
-        var workingFolder;
-        if (this.props.root) {
-            workingFolder = (
-                <span className="working-folder">(working folder: {this.props.root})</span>
-            );
-        }
-        return (
-            <div className="tree-node-folder">
-                <i className="fa fa-folder fa-fw" />
-                {label} {workingFolder}
-                <div className="files">
-                    {_.map(folder.children, this.renderChild)}
-                </div>
-            </div>
-        );
-    },
+  /**
+   * Render working folder at the root level
+   *
+   * @return {ReactElement|null}
+   */
+  function renderWorkingFolder() {
+    if (!root) {
+      return null;
+    }
+    return (
+      <span className="working-folder">
+        (working folder: {root})
+      </span>
+    );
+  }
 
-    /**
-     * Render child node
-     *
-     * @param  {Object} child
-     * @param  {Number} index
-     *
-     * @return {ReactElement}
-     */
-    renderChild: function(child, index) {
-        if (child.children) {
-            var props = {
-                key: index,
-                folder: child,
-                onSelect: this.props.onSelect,
-            };
-            return <TreeNodeFolder {...props} />;
-        } else {
-            var props = {
-                key: index,
-                file: child,
-                onSelect: this.props.onSelect,
-            };
-            return <TreeNodeFile {...props} />;
-        }
-    },
-});
+  /**
+   * Render child node
+   *
+   * @param  {Object} child
+   * @param  {Number} index
+   *
+   * @return {ReactElement}
+   */
+  function renderChild(child, index) {
+    if (child.children) {
+      const props = { folder: child, onSelect };
+      return <TreeNodeFolder key={index} {...props} />;
+    } else {
+      const props = { file: child, onSelect };
+      return <TreeNodeFile key={index} {...props} />;
+    }
+  }
+}
